requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
httpx==0.27.0
anthropic==0.28.0
apscheduler==3.10.4
jinja2==3.1.4
python-dotenv==1.0.1
cloudinary==1.40.0
reportlab==4.2.0
pypdf==4.2.0
PyYAML==6.0.2
